requirements.txt
asgiref==3.9.1
attrs==25.4.0
boto3==1.40.1
botocore==1.40.1
cachetools==6.2.2
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-resized==1.0.3
django-storages==1.14.6
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
google-auth==2.43.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.4
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.15.0
twilio==9.3.0
sendgrid==6.11.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
